Refactored out, this was like this before because of import cycles
