+++ infrastructure/permission/validPermissions.go
@@ -1,0 +1,94 @@
+package permission
+
+import "github.com/ystv/web-auth/permission/permissions"
+
+func GetValidPermissions(p permissions.Permissions) (m map[string]bool) {
+	m = make(map[string]bool)
+	if p.GetString() == permissions.MenuDisabled.GetString() {
+		m[p.GetString()] = true
+		return
+	}
+
+	m[p.GetString()] = true
+
+	switch p {
+	case permissions.ManageMembersAdmin:
+	case permissions.KeyCardAccess:
+	case permissions.BookingsAdmin:
+	case permissions.CalendarAdmin:
+	case permissions.CMSAdmin:
+	case permissions.Cobra:
+	case permissions.Director:
+	case permissions.EditNetStats:
+	case permissions.EmailEveryone:
+	case permissions.EquipmentAdmin:
+	case permissions.HiresAdmin:
+	case permissions.Inform:
+	case permissions.KeyListManage:
+	case permissions.MailingListAdmin:
+	case permissions.OfficerReports:
+	case permissions.Streamer:
+	case permissions.TechieTodo:
+	case permissions.VideoStats:
+	case permissions.WatchAdmin:
+		break
+	case permissions.ManageMembersMembersList:
+	case permissions.ManageMembersMembersAdd:
+		m[permissions.ManageMembersMembersAdmin.GetString()] = true
+	case permissions.ManageMembersPermissions:
+	case permissions.ManageMembersMicsKeyList:
+	case permissions.ManageMembersMiscUnpaidList:
+	case permissions.ManageMembersOfficers:
+	case permissions.ManageMembersGroup:
+	case permissions.ManageMembersMembersAdmin:
+		m[permissions.ManageMembersAdmin.GetString()] = true
+		break
+	case permissions.EmailAccess:
+	case permissions.EmailAlumni:
+	case permissions.EmailOfficers:
+		m[permissions.EmailEveryone.GetString()] = true
+		break
+	case permissions.CalendarSocialCreator:
+		m[permissions.CalendarSocialAdmin.GetString()] = true
+	case permissions.CalendarSocialAdmin:
+		m[permissions.CalendarAdmin.GetString()] = true
+		break
+	case permissions.CalendarShowCreator:
+		m[permissions.CalendarShowAdmin.GetString()] = true
+	case permissions.CalendarShowAdmin:
+		m[permissions.CalendarAdmin.GetString()] = true
+		break
+	case permissions.CalendarMeetingCreator:
+		m[permissions.CalendarMeetingAdmin.GetString()] = true
+	case permissions.CalendarMeetingAdmin:
+		m[permissions.CalendarAdmin.GetString()] = true
+		break
+	case permissions.CMSNewsItemCreator:
+		m[permissions.CMSNewsItemAdmin.GetString()] = true
+	case permissions.CMSNewsItemAdmin:
+		m[permissions.CMSNewsAdmin.GetString()] = true
+	case permissions.CMSEndboardAdmin:
+	case permissions.CMSView:
+	case permissions.CMSPermalinkAdmin:
+	case permissions.CMSNewsAdmin:
+		m[permissions.CMSAdmin.GetString()] = true
+		break
+	case permissions.CMSNewsCreator:
+		m[permissions.CMSNewsAdmin.GetString()] = true
+		m[permissions.CMSAdmin.GetString()] = true
+		break
+	case permissions.CMSPageCreator:
+		m[permissions.CMSPageAdmin.GetString()] = true
+	case permissions.CMSPageAdmin:
+		m[permissions.CMSAdmin.GetString()] = true
+		break
+	case permissions.CMSSlideshowCreator:
+		m[permissions.CMSSlideshowAdmin.GetString()] = true
+	case permissions.CMSSlideshowAdmin:
+		m[permissions.CMSAdmin.GetString()] = true
+		break
+	}
+
+	m[permissions.SuperUser.GetString()] = true
+	return
+}
